refactor: split notification handlers by protocol domain
